python/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
certifi==2023.7.22
click==8.1.7
fastapi==0.103.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pydantic==2.4.1
pydantic_core==2.10.1
python-telegram-bot==20.5
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzlocal==5.0.1
uvicorn==0.23.2
